Move to Feeds:: namespace
